requirements.txt
asgiref==3.3.1
boto==2.49.0
boto3==1.16.52
botocore==1.19.52
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
defusedxml==0.7.0rc1
Django==3.1.5
django-cors-headers==3.6.0
django-rest-auth==0.9.5
django-storages==1.11.1
django-templated-mail==1.1.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
idna==2.10
importlib-metadata==1.7.0
itypes==1.2.0
Jinja2==2.11.2
jmespath==0.10.0
MarkupSafe==1.1.1
oauthlib==3.1.0
Pillow==8.0.1
psycopg2==2.8.6
pycparser==2.20
PyJWT==2.0.0
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.5
requests==2.25.1
requests-oauthlib==1.3.0
s3transfer==0.3.4
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.0.2
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.2
zipp==3.4.0
